Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.enitties.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DataBaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service // Anotation utilizado para registrar serviços
@@ -30,8 +32,15 @@ public class UserService {
 		return repository.save(obj);
 	}
 
-	public void delete(Long id) {
-		repository.deleteById(id);
+	public void delete(Long id){
+	    try {
+	        if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+	        repository.deleteById(id);
+	    } catch (ResourceNotFoundException e){
+	        throw new ResourceNotFoundException(id);
+	    } catch(DataIntegrityViolationException e) {
+	    	throw new DataBaseException(e.getMessage());
+	    }
 	}
 
 	public User update(Long id, User obj) {
